Пытаемся парсить

+++ price_changer/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Product_from_wb(models.Model):
     prod_art_from_wb = models.CharField(max_length=255, unique=True, verbose_name='Артикул товара с WB')
-    price_with_discount_wb = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена с учетом скидки от кошелька')
+    price_with_discount_wb = models.IntegerField(null=True, blank=True, verbose_name='Цена с учетом скидки от кошелька')
 
     class Meta:
         verbose_name = "Товар с ВБ"
@@ -13,14 +13,14 @@ class Product_from_wb(models.Model):
     
 class Product_from_ozon(models.Model):
     prod_art_from_ozon = models.CharField(max_length=255, unique=True, verbose_name='Артикул товара с Ozon')
-    price_with_discount_ozon = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена с учетом скидки от кошелька')
-    price_ozon_s_be = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена, которая должна быть на Ozon')
-    price_with_co_invest = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена с соинвестом')
-    price_without_co_invest = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена без соинвеста')
-    discount_co_invest = models.IntegerField(null=True, blank=True, verbose_name='Скидка соинвеста')
-    discount_ozon_with_wallet = models.IntegerField(null=True, blank=True, verbose_name='Скидка кошелька Ozon')
-    price_ozon_s_be_with_wallet = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена, которая должна быть с учетом скидки кошелька Ozon')
-    price_ozon_s_be_with_co_invest = models.DecimalField(max_digits=10, decimal_places=0, null=True, blank=True, verbose_name='Цена, которая должна быть с учетом соинвеста')
+    price_with_discount_ozon = models.IntegerField(null=True, blank=True, verbose_name='Цена с учетом скидки от кошелька')
+    price_ozon_s_be = models.IntegerField(null=True, blank=True, verbose_name='Цена, которая должна быть на Ozon')
+    price_with_co_invest = models.IntegerField(null=True, blank=True, verbose_name='Цена с соинвестом')
+    price_without_co_invest = models.IntegerField(null=True, blank=True, verbose_name='Цена без соинвеста')
+    discount_co_invest = models.FloatField(null=True, blank=True, verbose_name='Скидка соинвеста')
+    discount_ozon_with_wallet = models.FloatField(null=True, blank=True, verbose_name='Скидка кошелька Ozon')
+    price_ozon_s_be_with_wallet = models.IntegerField(null=True, blank=True, verbose_name='Цена, которая должна быть с учетом скидки кошелька Ozon')
+    price_ozon_s_be_with_co_invest = models.IntegerField(null=True, blank=True, verbose_name='Цена, которая должна быть с учетом соинвеста')
 
     class Meta:
         verbose_name = "Товар с Озона"
